in progress

+++ Source/SmartHub/SmartHub.UWP.Plugins.Wemos/Core/WemosLineValueType.cs
@@ -5,7 +5,7 @@
         String,
         Boolean,
         Decimal,
-        Double,
+        Float,
         Custom
     }
 }
